Use a shared extension for LSP boilerplate
Handling the `Peer` directly within tests has gone a little smoother
than going through the `LanguageServer` interface. Pull out a shared
extension method to add the lifecycle methods that are required for
most or all tests. Add optional named arguments to override the defaults
for file lifecycle methods.

Replace the `MockLanguageServer`, which was never used as a mock, with
the same approach used in the tests.
