--- test/integration/lib/mock_lsp.dart
@@ -1,3 +1,0 @@
-import 'package:lsp/lsp.dart';
-
-class MockLanguageServer extends LanguageServer {}
