+++ test/integration/test/edit_test.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:_test/stub_lsp.dart';
 import 'package:_test/vim_remote.dart';
 import 'package:test/test.dart';
 import 'package:json_rpc_2/json_rpc_2.dart';
@@ -81,19 +82,4 @@ void main() {
     await Future.delayed(const Duration(milliseconds: 100));
     expect(await vim.expr(r'getline(1, "$")'), 'bar\nbar\n');
   });
-}
-
-extension LSP on Peer {
-  void registerLifecycleMethods(Map<String, dynamic> capabilities) {
-    registerMethod('initialize', (_) {
-      return {'capabilities': capabilities};
-    });
-    registerMethod('initialized', (_) {});
-    registerMethod('textDocument/didOpen', (_) {});
-    registerMethod('textDocument/didChange', (_) {});
-    registerMethod('shutdown', (_) {});
-    registerMethod('exit', (_) {
-      close();
-    });
-  }
 }
